i,src/Hello: Interpreter run script and make Hello work in it
The run script sets CWD so that we are using our stack.yaml (and thus have
available the packages/libraries it specifies as dependencies).

For the moment we add a UnicodeSyntax compiler directive to Hello.hs itself
because the interpreter's not reading the `default-extensions` list from
package.yaml. We need to figure out how to do that. If there's nothing
built-in, we may need just to read that list from the file and have our
interpreter startup script add the arguments to the clashi command line or
something like that.