package.yaml: Change main to specify module and definition
It used to specify the filename; it's not clear what exactly that does
(if anything) beyond including the file in the list of things to be
compiled. Despite what much documentation says, it's still possible to
specify an arbitrary module name and top-level definition as the entry
point. (Details are documented in the `Meet` file.)